fix(time): Fix responsibility from 'MoonPhase.js' to 'App.js'
